Update node.js
Fix issue where if a node is selected and the page is refreshed, the default node information is displayed instead of the selected node.

The page will now display the node's information if the selected node name is still in the url hash.  If not, it will display the default node information.